+++ nodes/Domus/methods/index.ts
@@ -1,0 +1,7 @@
+export {
+	searchBusinessTypes,
+	searchCities,
+	searchNeighborhoods,
+	searchPropertyTypes,
+	searchZones,
+} from './listSearch';